workers/pdf/pyproject.toml
[project]
name = "blobforge-pdf-worker"
version = "0.1.0"
description = "PDF to Markdown conversion worker for BlobForge"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "marker-pdf>=1.0.0",
]

[project.scripts]
blobforge-pdf-worker = "worker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
